Renamed and fixed tests
